pyproject.toml
[project]
name = "imagine-watcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "sse-starlette>=2.3.3",
    "uvicorn>=0.34.2",
]
